SHA-1: 793142870229453509a07c2a06dce463c6980140 正式添加对js脚本的支持（QScriptEngine）

+++ src/connectiondialog.cpp
@@ -34,6 +34,6 @@ void ConnectionDialog::on_connectButton_clicked()
 {
 	Config.UserName = ui->nameLineEdit->text();
 	Config.HostAddress = ui->hostLineEdit->text();
-	Config.Port = ui->portLineEdit->text().toInt();
+	Config.Port = ui->portLineEdit->text().toUShort();
 	accept();
 }
